+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 * fix `LiteralValueComparisonProcessor`: fix handling for NaN values
+* fix `LiteralValueComparisonProcessor`: fix deduplicated count measure
+* fix `ResourceValueComparisonProcessor`: fix deduplicated count measure
 
 ## [1.2.0] - 2023-03-09